Split flow class from the init module of the plugin
